+++ features/transaction/sellerorder/src/main/java/com/tokopedia/sellerorder/list/presentation/fragments/SomListFragment.kt
@@ -2634,7 +2634,7 @@ open class SomListFragment : BaseListFragment<Visitable<SomListAdapterTypeFactor
             mutableSetOf()
         } else {
             somListSortFilterTab?.addCounter(1)
-            mutableSetOf(filterOrderType)
+            mutableSetOf(filterOrderType.toLong())
         }
         setDefaultSortByValue()
         viewModel.setOrderTypeFilter(orderTypes)
